requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
docker-pycreds==0.4.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
grpcio==1.69.0
gym==0.21.0
idna==3.10
imageio==2.36.1
importlib-metadata==4.13.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
mujoco-py==2.1.2.14
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyDOE==0.3.8
pygame==2.6.1
pyglet==1.5.27
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.2
stable-baselines3==1.6.2
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.1.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wandb==0.19.2
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
# If MuJoCo raises errors
# importlib-metadata==4.13.0
# gym==0.21.0
# stable-baselines3[extra]==1.7.0